Update main function signatures to use 'void' for no parameters across multiple files
- Changed main function declarations from 'int main()' to 'int main(void)' to adhere to modern C standards and avoid compiler warnings.
- Updated function prototypes for skip_blanks and skip_comments to explicitly declare no parameters.
- Ensured consistent use of data types in loop variables to prevent sign comparison warnings.

+++ chapter_5/exercise_5_10/expr.c
@@ -14,7 +14,9 @@ void push(float element);
 int main(int argc, char *argv[]) {
     char Error = 0;
 
-    for (size_t i = 1; i < argc; ++i) {
+    // Using int instead of size_t for loop variable to match signed argc
+    // This prevents sign comparison warnings
+    for (int i = 1; i < argc; ++i) {
         float number = atof(argv[i]);
 
         if (number || strcmp(argv[i], "0") == 0) {
